Restructure folder for pip
